Keep the change inside the ticket's granted paths
The verifier caught four changed files with no grant behind them, and both
excursions were avoidable rather than necessary.

`scripts/cutcheck_contract.py` goes back to its baseline. The spec said
cutcheck imports the detector from the syntax owner, and `cutcheck_search`
is the granted file, so it reads `tickets_format` directly through the same
two-arm import every script here uses. Nothing else in cutcheck moves.

The lint cases fold into `tests/test_tickets_lint.py` itself. A case package
would have been three more files outside the scope for a shard whose cases
are all one subject.
