--- tests/test_tickets_lint_cases/__init__.py
@@ -1,1 +1,0 @@
-"""Cases for `tickets.py lint` and the tables it shares with dispatch."""
